requirements.txt
numpy==1.13.3
Pillow==3.1.2
